have one working collapisble container.  Need to move on to form and can figure out others later/stretch goal
