+++ src/CSS/Collapsible.css
@@ -1,0 +1,8 @@
+.Collapsible__trigger {
+    display: block;
+    font-weight: 400;
+    text-decoration: none;
+    position: relative;
+    border: 1px solid white;
+    padding: 10px;
+}
