+++ src/CSS/index.css
@@ -10,6 +10,6 @@ html {
   background-size: cover
 }
 
-#box {
+#main-box {
   margin-top: 2%
 }
